+++ CleanArchitecture/Source/InteractorProtocol.swift
@@ -3,7 +3,7 @@ import Foundation
 public protocol InteractorProtocol:AnyObject {
     associatedtype Router
     var router:Router? { get set }
-    var presenter:InteractorDelegateProtocol? { get set }
+    var presenter:InteractorDelegate? { get set }
     
     init()
     func didLoad()