+++ CleanArchitecture/Tests/MockInteractorProtocol.swift
@@ -3,7 +3,7 @@ import CleanArchitecture
 
 class MockInteractorProtocol:InteractorProtocol {
     var onDidLoad:(() -> Void)?
-    var presenter:InteractorDelegateProtocol?
+    var presenter:InteractorDelegate?
     var router:MockTransitionProtocol?
     
     required init() { }